Fix issue with tmux + find + stty (from coreutils)
Intermittently, but often, the SelectaFile and related commands failed
due to the following error: "stty: 'standard input': unable to perform
all requested operations". This only happens when using stty from GNU
coreutils along with find. The thing that seems to *work* is to separate
the find and stty calls so that they aren't part of the same pipeline.